+++ src/components/bottomSheetModalProvider/types.d.ts
@@ -4,7 +4,6 @@ import type { BottomSheetModalPrivateMethods } from '../bottomSheetModal';
 export interface BottomSheetModalRef {
   key: string;
   ref: RefObject<BottomSheetModalPrivateMethods>;
-  willUnmount: boolean;
 }
 
 export interface BottomSheetModalProviderProps {